getHexFloats() refactored to Utils
